// client/typescript/src/_generated/runtime/machine/entry.ts
/* generated by destack_client_generate, do not edit */

import { BinaryReader, BinaryWriter, Json, jsonField, jsonObject, jsonString } from "../../../protocol/serde.js";

/** User-facing runtime entrypoint name. */
export type Entry = {
    /** The fully qualified entry name. */
    readonly name: string;
};

export const Entry = {
    /** Encode this value. */
    encode(writer: BinaryWriter, value: Entry): void {
        encodeEntry(writer, value);
    },

    /** Decode one Entry. */
    decode(reader: BinaryReader): Entry {
        return decodeEntry(reader);
    },

    /** Return this value as JSON. */
    toJson(value: Entry): Json {
        return toJsonEntry(value);
    },

    /** Return one Entry from one JSON value. */
    fromJson(value: Json): Entry {
        return fromJsonEntry(value);
    },
};

/** Encode one Entry. */
export function encodeEntry(writer: BinaryWriter, value: Entry): void {
    writer.writeString(value.name);
}

/** Decode one Entry. */
export function decodeEntry(reader: BinaryReader): Entry {
    const name = reader.readString();

    return {
        name,
    };
}

/** Return one JSON value for one Entry. */
export function toJsonEntry(value: Entry): Json {
    return {
        name: value.name,
    };
}

/** Return one Entry from one JSON value. */
export function fromJsonEntry(value: Json): Entry {
    const object = jsonObject(value);

    return {
        name: jsonString(jsonField(object, "name")),
    };
}